pyproject.toml
[project]
name = "bullseye"
version = "0.0.1"
description = "Script to organize wallpapers based on labels"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clip",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.1",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
]
